requirements.txt
requests
python-dotenv
apprise
